requirements.txt
flask
SQLAlchemy
python-dotenv
flask-SQLAlchemy
flask-bcrypt
flask-session
redis